requirements.txt
numpy==1.26.2
tqdm==4.66.1
spacy==3.7.2
nltk==3.8.1
webvtt-py==0.4.6
yt_dlp==2023.11.16
fastapi==0.108.0
uvicorn[standard]==0.24.0.post1
SQLAlchemy==2.0.25
aiosqlite==0.19.0
gunicorn==21.2.0
networkx==3.2.1
pytest==7.4.4
pytest-asyncio==0.23.3
beautifulsoup4==4.12.2
httpx==0.25.2
scipy==1.11.4
